Perform request against PunkApi with Retrofit.
 * ADD: Retrofit library to perform the request against the api.
 * ADD: Call in the activity to perform a request to get a random beer
 * FIX: Error with the BeerDTO, changed some fields from int to double.
 * ADD: Missing internet permissions on the manifest

+++ app/src/main/java/com/p4r4d0x/punkapibeerviewer/model/BeerDataClasses.kt
@@ -65,7 +65,7 @@ data class BeerDTO(
     @SerializedName("first_brewed")
     val firstBrewed: String = "",
     @SerializedName("attenuation_level")
-    val attenuationLevel: Int = 0,
+    val attenuationLevel: Double = 0.0,
     @SerializedName("method")
     val method: Method,
     @SerializedName("target_og")
@@ -81,7 +81,7 @@ data class BeerDTO(
     @SerializedName("target_fg")
     val targetFg: Int = 0,
     @SerializedName("srm")
-    val srm: Int = 0,
+    val srm: Double = 0.0,
     @SerializedName("volume")
     val volume: Volume,
     @SerializedName("contributed_by")
